scripts/requirements.txt
# Google API dependencies for export_to_google_docs.py
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
